implement virtual backends
Adds the ability to create virtual backends. Virtual backends share a
single location but route to different backends depending on the rules
defined by the VirtualBackend type.

When the list of upstreams and servers are retrieved, we then call
mergeVirtualBackends which iterates through the paths of every ingress
and checks if the backend supporting the path is a VirtualBackend. If
so, we then iterate through the map of servers and find the real backend
that the VirtualBackend should fall under. Once found, the
VirtualBackend is embedded in the list of VirtualBackends for the real
backend and the VirtualBackend's location is deleted.

If no matching real backend for a VirtualBackend is found, then the
VirtualBackend and locations associated with it are deleted as they
cannot be backed by any server.